+++ tests/src/test/java/com/orientechnologies/orient/test/database/auto/SQLFunctionsTest.java
@@ -210,7 +210,7 @@ public class SQLFunctionsTest {
 
 		Assert.assertTrue(result.size() != 0);
 		for (ODocument d : result) {
-			Assert.assertTrue((Long) d.field("nr") <= 1000);
+			Assert.assertTrue((Integer) d.field("nr") <= 1000);
 		}
 
 		OSQLEngine.getInstance().unregisterInlineFunction("bigger");
